src/requirements.txt
Django==3.1.7
requests==2.25.1
Markdown==3.2.1
pandas==1.2.3
djangorestframework==3.12.2
django-filter==2.4.0
